refactor: Post관련 테스트 분리
