+++ backend/src/test/java/com/backjoongwon/cvi/post/application/PostCommentServiceTest.java
@@ -1,0 +1,119 @@
+package com.backjoongwon.cvi.post.application;
+
+import com.backjoongwon.cvi.comment.domain.Comment;
+import com.backjoongwon.cvi.comment.dto.CommentRequest;
+import com.backjoongwon.cvi.comment.dto.CommentResponse;
+import com.backjoongwon.cvi.common.exception.NotFoundException;
+import com.backjoongwon.cvi.common.exception.UnAuthorizedException;
+import com.backjoongwon.cvi.post.domain.Post;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+@DisplayName("게시글-댓글 비즈니스 흐름 테스트")
+public class PostCommentServiceTest extends InitPostServiceTest {
+
+    @DisplayName("댓글 생성 - 성공")
+    @Test
+    void createComment() {
+        //given
+        //when
+        CommentRequest newCommentRequest = new CommentRequest("새로운 댓글 내용");
+        CommentResponse commentResponse = postService.createComment(post1.getId(), optionalAnotherUser, newCommentRequest);
+        Post foundPost = postRepository.findWithCommentsById(post1.getId())
+                .orElseThrow(() -> new NotFoundException("해당 id의 게시글이 없습니다."));
+        //then
+        assertThat(foundPost.getCommentsAsList()).extracting("id").contains(commentResponse.getId());
+    }
+
+    @DisplayName("댓글 생성 - 실패 - 비회원 댓글 작성 시도")
+    @Test
+    void createCommentWhenNotLoginUser() {
+        //given
+        //when
+        //then
+        assertThatThrownBy(() -> postService.createComment(post2.getId(), Optional.empty(), commentRequest))
+                .isInstanceOf(UnAuthorizedException.class);
+    }
+
+    @DisplayName("댓글 수정 - 성공")
+    @Test
+    void updateComment() {
+        //given
+        CommentResponse commentResponse = postService.createComment(post2.getId(), optionalUser, commentRequest);
+        CommentRequest updateRequest = new CommentRequest("저녁 술 ㄱ?");
+        //when
+        postService.updateComment(post2.getId(), commentResponse.getId(), optionalUser, updateRequest);
+        Comment comment = commentRepository.findById(commentResponse.getId()).get();
+        //then
+        assertThat(comment.getContent()).isEqualTo(updateRequest.getContent());
+    }
+
+    @DisplayName("댓글 수정 - 실패 - 댓글을 찾을 수 없음")
+    @Test
+    void updateCommentWhenNoComment() {
+        //given
+        CommentRequest updateRequest = new CommentRequest("저녁 술 ㄱ?");
+        //when
+        //then
+        assertThatThrownBy(() -> postService.updateComment(post2.getId(), 0L, optionalUser, updateRequest))
+                .isInstanceOf(NotFoundException.class)
+                .hasMessage("찾을 수 없는 댓글입니다.");
+    }
+
+    @DisplayName("댓글 수정 - 실패 - 댓글 작성자가 아님")
+    @Test
+    void updateCommentWhenWrongUser() {
+        //given
+        CommentResponse commentResponse = postService.createComment(post2.getId(), optionalUser, commentRequest);
+        CommentRequest updateRequest = new CommentRequest("저녁 술 ㄱ?");
+        //when
+        //then
+        assertThatThrownBy(() -> postService.updateComment(post2.getId(), commentResponse.getId(), optionalAnotherUser, updateRequest))
+                .isInstanceOf(UnAuthorizedException.class)
+                .hasMessage("다른 사용자의 게시글은 수정할 수 없습니다.");
+    }
+
+    @DisplayName("댓글 삭제 - 성공")
+    @Test
+    void deleteComment() {
+        //given
+        CommentResponse commentResponse = postService.createComment(post2.getId(), optionalUser, commentRequest);
+        //when
+        postService.deleteComment(post2.getId(), commentResponse.getId(), optionalUser);
+        postRepository.flush();
+
+        Optional<Comment> foundComment = commentRepository.findById(commentResponse.getId());
+        Post foundPost = postRepository.findById(this.post2.getId()).get();
+        //then
+        assertThat(foundComment).isEmpty();
+        assertThat(foundPost.getCommentsAsList()).extracting("id").doesNotContain(commentResponse.getId());
+    }
+
+    @DisplayName("댓글 삭제 - 실패 - 댓글을 찾을 수 없음")
+    @Test
+    void deleteCommentWhenNoComment() {
+        //given
+        //when
+        //then
+        assertThatThrownBy(() -> postService.deleteComment(post2.getId(), 0L, optionalUser))
+                .isInstanceOf(NotFoundException.class)
+                .hasMessage("찾을 수 없는 댓글입니다.");
+    }
+
+    @DisplayName("댓글 삭제 - 실패 - 댓글 작성자가 아님")
+    @Test
+    void deleteCommentWhenWrongUser() {
+        //given
+        CommentResponse commentResponse = postService.createComment(post2.getId(), optionalUser, commentRequest);
+        //when
+        //then
+        assertThatThrownBy(() -> postService.deleteComment(post2.getId(), commentResponse.getId(), optionalAnotherUser))
+                .isInstanceOf(UnAuthorizedException.class)
+                .hasMessage("다른 사용자의 게시글은 삭제할 수 없습니다.");
+    }
+}
